refactor: 型をファイル分割した
